requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
codenamize==1.2.3
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
gradio==4.31.4
gradio_client==0.16.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python_avatars==1.4.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.4
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.0
semantic-version==2.10.0
sentence-transformers==2.7.0
sentiment3d @ git+https://github.com/compasspathways/Sentiment3D.git@b29fc1d1d5b7b05cc2941a3da7d18ca14df1144f
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
statsmodels==0.14.2
sympy==1.12
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.3
